Refactor edit message
